backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.114.1
h11==0.14.0
idna==3.8
lxml==5.3.0
marshmallow==3.23.0
packaging==24.1
pillow==10.4.0
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.10.1
python-dotenv==1.0.1
python-pptx==1.0.2
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
XlsxWriter==3.2.0
